# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NovelSync Story Data API
  version: 0.1.0
  description: PostgreSQL system-of-record API. Browser requests use Firebase ID tokens.
servers:
  - url: https://story-data.example.invalid
paths:
...
  /v1/stories:
    get:
      security: [{ firebaseBearer: [] }]
      responses: { '200': { description: Owner stories } }
    post:
      security: [{ firebaseBearer: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StoryInput' } } } }
      responses: { '201': { description: Created } }
  /v1/public/stories:
    get:
      summary: List published stories
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 24 } }
      responses: { '200': { description: Cursor-paginated public story summaries } }
  /v1/public/profiles:
    get:
      summary: Search, list recent, or batch-read public member profiles
      parameters:
        - { name: query, in: query, schema: { type: string } }
        - { name: ids, in: query, schema: { type: string, description: Comma-separated user IDs; maximum 50 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses: { '200': { description: Public profiles } }
...
  /v1/profiles/me:
    get:
      security: [{ firebaseBearer: [] }]
      responses: { '200': { description: Current user's public profile } }
    put:
      security: [{ firebaseBearer: [] }]
      responses: { '201': { description: Created or replaced profile }, '409': { description: Username is taken } }
    patch:
      security: [{ firebaseBearer: [] }]
      responses: { '200': { description: Updated profile }, '409': { description: Username is taken } }
  /v1/me/reading-history:
    get:
      security: [{ firebaseBearer: [] }]
      parameters: [{ name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 5 } }]
      responses: { '200': { description: Recent private reading history } }
    delete:
      security: [{ firebaseBearer: [] }]
      responses: { '204': { description: Cleared } }
...
  /v1/competitions:
    get:
      summary: List published competitions, newest first
      description: >-
        Returns a bare array. When more pages remain, the continuation token is
        in the X-Next-Cursor response header; pass it back as ?cursor=.
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Competitions
          headers:
            X-Next-Cursor:
              schema: { type: string }
              description: Present only when a further page exists.
        '400': { description: limit is not a positive integer }
        '422': { description: cursor is malformed }
...
components:
  securitySchemes:
    firebaseBearer: { type: http, scheme: bearer, bearerFormat: Firebase-ID-token }
  schemas:
    StoryInput:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 500 }
        description: { type: string }
        authorName: { type: string }
        published: { type: boolean }